+++ src/Main.java
@@ -3,10 +3,10 @@ public class Main {
     public static void main(String[] args)
     {
         BonusMilesService service = new BonusMilesService();
-        float price = 0;
+        int price = 0;
         Scanner in = new Scanner(System.in);
         System.out.print ("Введите цену билета \n");
-        price = in.nextFloat();
+        price = in.nextInt();
         System.out.print ("Вам начислено бесплатных миль \n" + service.calculate (price));
     }
 }
